--- dev/20190701/k2works/api/rental/src/main/java/generics/package-info.java
@@ -1,1 +1,0 @@
-package generics;
