--- dev/20190701/k2works/api/rental/src/main/java/rental/charge/package-info.java
@@ -1,1 +1,0 @@
-package rental.charge;
