+++ dev/20190701/k2works/api/rental/src/main/java/rental/domain/generics/Accumulate.java
@@ -1,4 +1,4 @@
-package generics;
+package rental.domain.generics;
 
 public interface Accumulate<T> {
   T add(T other);
