+++ dev/20190701/k2works/api/rental/src/main/java/rental/domain/generics/Total.java
@@ -1,4 +1,4 @@
-package generics;
+package rental.domain.generics;
 
 import java.util.List;
 import java.util.function.Function;
