+++ dev/20190701/k2works/api/rental/src/main/java/rental/domain/model/DaysRented.java
@@ -1,4 +1,4 @@
-package rental;
+package rental.domain.model;
 
 public class DaysRented {
   int value;
