+++ dev/20190701/k2works/api/rental/src/main/java/rental/domain/model/charge/package-info.java
@@ -1,0 +1,1 @@
+package rental.domain.model.charge;
