+++ dev/20190701/k2works/api/rental/src/main/java/rental/domain/model/customer/Customer.java
@@ -1,4 +1,4 @@
-package rental.customer;
+package rental.domain.model.customer;
 
 public class Customer {
   String name;
