+++ dev/20190701/k2works/api/rental/src/main/java/rental/domain/model/movie/Movie.java
@@ -1,4 +1,4 @@
-package rental.movie;
+package rental.domain.model.movie;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
